+++ previous-versions/accounting/src/codataccounting/bill_payments.py
@@ -30,7 +30,7 @@ class BillPayments:
         
         url = utils.generate_url(operations.CreateBillPaymentRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/billPayments', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "bill_payment", True, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.CreateBillPaymentRequest, "bill_payment", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreateBillPaymentRequest, request)