+++ previous-versions/accounting/src/codataccounting/models/errors/errormessage.py
@@ -27,4 +27,4 @@ class ErrorMessage(Exception):
     
 
     def __str__(self) -> str:
-        return utils.marshal_json(self)
+        return utils.marshal_json(self, type(self))