+++ previous-versions/accounting/src/codataccounting/payments.py
@@ -30,7 +30,7 @@ class Payments:
         
         url = utils.generate_url(operations.CreatePaymentRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/payments', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "payment", True, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.CreatePaymentRequest, "payment", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreatePaymentRequest, request)