+++ previous-versions/accounting/src/codataccounting/suppliers.py
@@ -35,7 +35,7 @@ class Suppliers:
         
         url = utils.generate_url(operations.CreateSupplierRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/suppliers', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "supplier", True, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.CreateSupplierRequest, "supplier", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreateSupplierRequest, request)
@@ -472,7 +472,7 @@ class Suppliers:
         
         url = utils.generate_url(operations.UpdateSupplierRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/suppliers/{supplierId}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "supplier", True, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.UpdateSupplierRequest, "supplier", True, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.UpdateSupplierRequest, request)