+++ previous-versions/accounting/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.33.0] previous-versions/accounting
 ### Releases
 - [PyPI v0.33.0] https://pypi.org/project/codat-accounting/0.33.0 - previous-versions/accounting
+
+## 2023-12-14 14:04:36
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] previous-versions/accounting
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-accounting/0.34.0 - previous-versions/accounting